requirements.txt
mysql-connector==2.2.9
numpy==1.22.3
Pillow==9.1.1
SQLAlchemy==1.4.36
